@@ -1,7 +1,10 @@
-const { ethers } = require("ethers");
-const { bufferToHex, keccak256 } = require("ethereumjs-util");
+import { ethers } from "ethers";
 
-const merkleTree = (tokenIds) => {
+const { keccak256 } = ethers.utils;
+
+type BufferElementPositionIndex = { [key: string]: number };
+
+export const merkleTree = (tokenIds: ethers.BigNumber[]) => {
   const elements = tokenIds
     .map((tokenId) =>
       Buffer.from(tokenId.toHexString().slice(2).padStart(64, "0"), "hex")
@@ -11,19 +14,22 @@ const merkleTree = (tokenIds) => {
       return idx === 0 || !arr[idx - 1].equals(el);
     });
 
-  const bufferElementPositionIndex = elements.reduce((memo, el, index) => {
-    memo[bufferToHex(el)] = index;
-    return memo;
-  }, {});
+  const bufferElementPositionIndex = elements.reduce(
+    (memo: BufferElementPositionIndex, el, index) => {
+      memo["0x" + el.toString("hex")] = index;
+      return memo;
+    },
+    {}
+  );
 
   // Create layers
   const layers = getLayers(elements);
 
-  const root = bufferToHex(layers[layers.length - 1][0]);
+  const root = "0x" + layers[layers.length - 1][0].toString("hex");
 
   const proofs = Object.fromEntries(
     elements.map((el) => [
-      ethers.BigNumber.from("0x" + el.toString("hex")).toString(),
+      ethers.BigNumber.from(el).toString(),
       getHexProof(el, bufferElementPositionIndex, layers),
     ])
   );
@@ -39,13 +45,13 @@ const merkleTree = (tokenIds) => {
   };
 };
 
-const getLayers = (elements) => {
+const getLayers = (elements: Buffer[]) => {
   if (elements.length === 0) {
     throw new Error("empty tree");
   }
 
   const layers = [];
-  layers.push(elements.map((el) => keccak256(el)));
+  layers.push(elements.map((el) => Buffer.from(keccak256(el).slice(2), "hex")));
 
   // Get next layer until we reach the root
   while (layers[layers.length - 1].length > 1) {
@@ -55,8 +61,8 @@ const getLayers = (elements) => {
   return layers;
 };
 
-const getNextLayer = (elements) => {
-  return elements.reduce((layer, el, idx, arr) => {
+const getNextLayer = (elements: Buffer[]) => {
+  return elements.reduce((layer: Buffer[], el, idx, arr) => {
     if (idx % 2 === 0) {
       // Hash the current element with its pair element
       layer.push(combinedHash(el, arr[idx + 1]));
@@ -66,7 +72,7 @@ const getNextLayer = (elements) => {
   }, []);
 };
 
-const combinedHash = (first, second) => {
+const combinedHash = (first: Buffer, second: Buffer) => {
   if (!first) {
     return second;
   }
@@ -74,17 +80,24 @@ const combinedHash = (first, second) => {
     return first;
   }
 
-  return keccak256(Buffer.concat([first, second].sort(Buffer.compare)));
+  return Buffer.from(
+    keccak256(Buffer.concat([first, second].sort(Buffer.compare))).slice(2),
+    "hex"
+  );
 };
 
-const getHexProof = (el, bufferElementPositionIndex, layers) => {
-  let idx = bufferElementPositionIndex[bufferToHex(el)];
+const getHexProof = (
+  el: Buffer,
+  bufferElementPositionIndex: BufferElementPositionIndex,
+  layers: Buffer[][]
+) => {
+  let idx = bufferElementPositionIndex["0x" + el.toString("hex")];
 
   if (typeof idx !== "number") {
     throw new Error("Element does not exist in Merkle tree");
   }
 
-  const proofBuffer = layers.reduce((proof, layer) => {
+  const proofBuffer = layers.reduce((proof: Buffer[], layer) => {
     const pairIdx = idx % 2 === 0 ? idx + 1 : idx - 1;
     const pairElement = pairIdx < layer.length ? layer[pairIdx] : null;
 
@@ -99,7 +112,3 @@ const getHexProof = (el, bufferElementPositionIndex, layers) => {
 
   return proofBuffer.map((el) => "0x" + el.toString("hex"));
 };
-
-module.exports = Object.freeze({
-  merkleTree,
-});
