@@ -1,17 +1,21 @@
 import { expect } from "chai";
-import { constants, Wallet } from "ethers";
-import {
+import { constants } from "ethers";
+import { hexZeroPad } from "ethers/lib/utils";
+import { network } from "hardhat";
+import { getScuffedContract } from "scuffed-abi";
+
+import { buildOrderStatus, getBasicOrderParameters } from "../utils/encoding";
+import { getWalletWithEther } from "../utils/faucet";
+import { seaportFixture } from "../utils/fixtures";
+
+import type {
   ConsiderationInterface,
   TestERC20,
   TestERC721,
 } from "../../typechain-types";
-import { buildOrderStatus, getBasicOrderParameters } from "../utils/encoding";
-import { seaportFixture, SeaportFixtures } from "../utils/fixtures";
-import { getWalletWithEther } from "../utils/impersonate";
-import { AdvancedOrder, ConsiderationItem } from "../utils/types";
-import { getScuffedContract } from "scuffed-abi";
-import { hexZeroPad } from "ethers/lib/utils";
-import { network } from "hardhat";
+import type { SeaportFixtures } from "../utils/fixtures";
+import type { AdvancedOrder, ConsiderationItem } from "../utils/types";
+import type { Wallet } from "ethers";
 
 const IS_FIXED = true;
 
@@ -19,16 +23,19 @@ describe("Additional recipients off by one error allows skipping second consider
   let alice: Wallet;
   let bob: Wallet;
   let carol: Wallet;
+
   let order: AdvancedOrder;
   let orderHash: string;
+
+  let marketplaceContract: ConsiderationInterface;
   let testERC20: TestERC20;
   let testERC721: TestERC721;
-  let marketplaceContract: ConsiderationInterface;
-  let mintAndApprove721: SeaportFixtures["mintAndApprove721"];
-  let mintAndApproveERC20: SeaportFixtures["mintAndApproveERC20"];
+
+  let createOrder: SeaportFixtures["createOrder"];
   let getTestItem20: SeaportFixtures["getTestItem20"];
   let getTestItem721: SeaportFixtures["getTestItem721"];
-  let createOrder: SeaportFixtures["createOrder"];
+  let mintAndApprove721: SeaportFixtures["mintAndApprove721"];
+  let mintAndApproveERC20: SeaportFixtures["mintAndApproveERC20"];
 
   after(async () => {
     await network.provider.request({
@@ -40,18 +47,21 @@ describe("Additional recipients off by one error allows skipping second consider
     alice = await getWalletWithEther();
     bob = await getWalletWithEther();
     carol = await getWalletWithEther();
+
     ({
-      mintAndApprove721,
-      mintAndApproveERC20,
-      marketplaceContract,
+      createOrder,
       getTestItem20,
       getTestItem721,
-      createOrder,
+      marketplaceContract,
+      mintAndApprove721,
+      mintAndApproveERC20,
       testERC20,
       testERC721,
     } = await seaportFixture(await getWalletWithEther()));
+
     // ERC721 with ID = 1
     await mintAndApprove721(alice, marketplaceContract.address, 1);
+
     // ERC20 with amount = 1100
     await mintAndApproveERC20(bob, marketplaceContract.address, 1100);
   });
@@ -87,9 +97,72 @@ describe("Additional recipients off by one error allows skipping second consider
     );
 
     // Bob validates the order
-    await expect(marketplaceContract.connect(bob).validate([order]))
-      .to.emit(marketplaceContract, "OrderValidated")
-      .withArgs(orderHash, alice.address, constants.AddressZero);
+    const tx = await marketplaceContract.connect(bob).validate([order]);
+
+    const receipt = await tx.wait();
+
+    expect(receipt.events?.length).to.equal(1);
+
+    const event = receipt.events && receipt.events[0];
+
+    expect(event?.event).to.equal("OrderValidated");
+
+    expect(event?.args?.orderHash).to.equal(orderHash);
+
+    const parameters = event && event.args && event.args.orderParameters;
+
+    expect(parameters.offerer).to.equal(order.parameters.offerer);
+    expect(parameters.zone).to.equal(order.parameters.zone);
+    expect(parameters.orderType).to.equal(order.parameters.orderType);
+    expect(parameters.startTime).to.equal(order.parameters.startTime);
+    expect(parameters.endTime).to.equal(order.parameters.endTime);
+    expect(parameters.zoneHash).to.equal(order.parameters.zoneHash);
+    expect(parameters.salt).to.equal(order.parameters.salt);
+    expect(parameters.conduitKey).to.equal(order.parameters.conduitKey);
+    expect(parameters.totalOriginalConsiderationItems).to.equal(
+      order.parameters.totalOriginalConsiderationItems
+    );
+    expect(parameters.totalOriginalConsiderationItems).to.equal(
+      parameters.consideration.length
+    );
+
+    expect(parameters.offer.length).to.equal(order.parameters.offer.length);
+    expect(parameters.consideration.length).to.equal(
+      order.parameters.consideration.length
+    );
+
+    for (let i = 0; i < parameters.offer.length; i++) {
+      const eventOffer = parameters.offer[i];
+      const suppliedOffer = order.parameters.offer[i];
+      expect(eventOffer.itemType).to.equal(suppliedOffer.itemType);
+      expect(eventOffer.token).to.equal(suppliedOffer.token);
+      expect(eventOffer.identifierOrCriteria).to.equal(
+        suppliedOffer.identifierOrCriteria
+      );
+      expect(eventOffer.startAmount).to.equal(suppliedOffer.startAmount);
+      expect(eventOffer.endAmount).to.equal(suppliedOffer.endAmount);
+    }
+
+    for (let i = 0; i < parameters.consideration.length; i++) {
+      const eventConsideration = parameters.consideration[i];
+      const suppliedConsideration = order.parameters.consideration[i];
+      expect(eventConsideration.itemType).to.equal(
+        suppliedConsideration.itemType
+      );
+      expect(eventConsideration.token).to.equal(suppliedConsideration.token);
+      expect(eventConsideration.identifierOrCriteria).to.equal(
+        suppliedConsideration.identifierOrCriteria
+      );
+      expect(eventConsideration.startAmount).to.equal(
+        suppliedConsideration.startAmount
+      );
+      expect(eventConsideration.endAmount).to.equal(
+        suppliedConsideration.endAmount
+      );
+      expect(eventConsideration.recipient).to.equal(
+        suppliedConsideration.recipient
+      );
+    }
 
     // OrderStatus is validated
     orderStatus = await marketplaceContract.getOrderStatus(orderHash);
@@ -150,8 +223,12 @@ describe("Additional recipients off by one error allows skipping second consider
           bob.sendTransaction({
             to: marketplaceContract.address,
             data: maliciousCallData,
+            gasLimit: 29_999_999,
           })
-        ).to.be.revertedWith("MissingOriginalConsiderationItems");
+        ).to.be.revertedWithCustomError(
+          marketplaceContract,
+          "MissingOriginalConsiderationItems"
+        );
       });
     }
   });
