@@ -1,15 +1,18 @@
 import { ethers } from "hardhat";
-import { Contract } from "ethers";
-import { JsonRpcSigner } from "@ethersproject/providers";
-import * as dotenv from "dotenv";
 
-dotenv.config();
+import type { JsonRpcSigner } from "@ethersproject/providers";
+import type { Contract, Wallet } from "ethers";
+
+import "dotenv/config";
 
 export const deployContract = async <C extends Contract>(
   name: string,
-  signer: JsonRpcSigner,
+  signer?: JsonRpcSigner | Wallet,
   ...args: any[]
 ): Promise<C> => {
+  if (!signer) {
+    signer = await ethers.provider.getSigner(0);
+  }
   const references = new Map<string, string>([
     ["Consideration", "ReferenceConsideration"],
     ["Conduit", "ReferenceConduit"],
@@ -18,7 +21,7 @@ export const deployContract = async <C extends Contract>(
 
   const nameWithReference =
     process.env.REFERENCE && references.has(name)
-      ? references.get(name) || name
+      ? references.get(name) ?? name
       : name;
 
   const f = await ethers.getContractFactory(nameWithReference, signer);
