@@ -1,18 +1,19 @@
+import { expect } from "chai";
 import { randomBytes as nodeRandomBytes } from "crypto";
-import { utils, BigNumber, constants, ContractTransaction } from "ethers";
+import { BigNumber, constants, utils } from "ethers";
 import { getAddress, keccak256, toUtf8Bytes } from "ethers/lib/utils";
-import {
+
+import type {
   BasicOrderParameters,
-  BigNumberish,
   ConsiderationItem,
   CriteriaResolver,
+  Fulfillment,
   FulfillmentComponent,
   OfferItem,
   Order,
   OrderComponents,
 } from "./types";
-
-export { BigNumberish };
+import type { BigNumberish, ContractTransaction } from "ethers";
 
 const SeededRNG = require("./seeded-rng");
 
@@ -26,8 +27,6 @@ if (GAS_REPORT_MODE) {
   randomBytes = (n: number) => nodeRandomBytes(n).toString("hex");
 }
 
-// const randomBytes
-
 export const randomHex = (bytes = 32) => `0x${randomBytes(bytes)}`;
 
 export const random128 = () => toBN(randomHex(16));
@@ -40,8 +39,8 @@ export const toHex = (n: BigNumberish, numBytes: number = 0) => {
     : typeof n === "string"
     ? hexRegex.test(n)
       ? n.replace(/0x/, "")
-      : (+n).toString(16)
-    : (+n).toString(16);
+      : Number(n).toString(16)
+    : Number(n).toString(16);
   return `0x${asHexString.padStart(numBytes * 2, "0")}`;
 };
 
@@ -70,9 +69,7 @@ export const convertSignatureToEIP2098 = (signature: string) => {
     return signature;
   }
 
-  if (signature.length !== 132) {
-    throw Error("invalid signature length (must be 64 or 65 bytes)");
-  }
+  expect(signature.length, "signature must be 64 or 65 bytes").to.eq(132);
 
   return utils.splitSignature(signature).compact;
 };
@@ -80,8 +77,8 @@ export const convertSignatureToEIP2098 = (signature: string) => {
 export const getBasicOrderParameters = (
   basicOrderRouteType: number,
   order: Order,
-  fulfillerConduitKey = false,
-  tips = []
+  fulfillerConduitKey: string | boolean = false,
+  tips: { amount: BigNumber; recipient: string }[] = []
 ): BasicOrderParameters => ({
   offerer: order.parameters.offerer,
   zone: order.parameters.zone,
@@ -102,7 +99,9 @@ export const getBasicOrderParameters = (
   ),
   signature: order.signature,
   offererConduitKey: order.parameters.conduitKey,
-  fulfillerConduitKey: toKey(fulfillerConduitKey),
+  fulfillerConduitKey: toKey(
+    typeof fulfillerConduitKey === "string" ? fulfillerConduitKey : 0
+  ),
   additionalRecipients: [
     ...order.parameters.consideration
       .slice(1)
@@ -189,10 +188,7 @@ export const toFulfillmentComponents = (
 export const toFulfillment = (
   offerArr: number[][],
   considerationsArr: number[][]
-): {
-  offerComponents: FulfillmentComponent[];
-  considerationComponents: FulfillmentComponent[];
-} => ({
+): Fulfillment => ({
   offerComponents: toFulfillmentComponents(offerArr),
   considerationComponents: toFulfillmentComponents(considerationsArr),
 });
@@ -322,8 +318,8 @@ export const getBasicOrderExecutions = (
         amount: offerItem.endAmount,
         recipient: fulfiller,
       },
-      offerer: offerer,
-      conduitKey: conduitKey,
+      offerer,
+      conduitKey,
     },
     {
       item: {
