@@ -1,15 +1,17 @@
-/* eslint-disable camelcase */
 import { expect } from "chai";
-import { constants, Wallet } from "ethers";
+import { constants } from "ethers";
 import { getCreate2Address, keccak256 } from "ethers/lib/utils";
 import hre, { ethers } from "hardhat";
-import {
-  ConduitControllerInterface,
-  ImmutableCreate2FactoryInterface,
-} from "../../../typechain-types";
+
 import { deployContract } from "../contracts";
 import { randomHex } from "../encoding";
-import { whileImpersonating } from "../impersonate";
+
+import type {
+  ConduitControllerInterface,
+  Conduit__factory,
+  ImmutableCreate2FactoryInterface,
+} from "../../../typechain-types";
+import type { Wallet } from "ethers";
 
 const deployConstants = require("../../../constants/constants");
 
@@ -18,10 +20,12 @@ export const conduitFixture = async (
   owner: Wallet
 ) => {
   let conduitController: ConduitControllerInterface;
-  let conduitImplementation: any;
+  let conduitImplementation: Conduit__factory;
   if (process.env.REFERENCE) {
-    conduitImplementation = await ethers.getContractFactory("ReferenceConduit");
-    conduitController = await deployContract("ConduitController", owner as any);
+    conduitImplementation = (await ethers.getContractFactory(
+      "ReferenceConduit"
+    )) as Conduit__factory;
+    conduitController = await deployContract("ConduitController", owner);
   } else {
     conduitImplementation = await ethers.getContractFactory("Conduit");
 
@@ -82,23 +86,29 @@ export const conduitFixture = async (
   const deployNewConduit = async (owner: Wallet, conduitKey?: string) => {
     // Create a conduit key with a random salt
     const assignedConduitKey =
-      conduitKey || owner.address + randomHex(12).slice(2);
+      conduitKey ?? owner.address + randomHex(12).slice(2);
 
     const { conduit: tempConduitAddress } = await conduitController.getConduit(
       assignedConduitKey
     );
 
-    await whileImpersonating(owner.address, ethers.provider, async () => {
+    if (!process.env.REFERENCE) {
       await expect(
         conduitController
           .connect(owner)
           .createConduit(assignedConduitKey, constants.AddressZero)
-      ).to.be.revertedWith("InvalidInitialOwner");
+      ).to.be.revertedWithCustomError(conduitController, "InvalidInitialOwner");
+    } else {
+      await expect(
+        conduitController
+          .connect(owner)
+          .createConduit(assignedConduitKey, constants.AddressZero)
+      ).to.be.reverted;
+    }
 
-      await conduitController
-        .connect(owner)
-        .createConduit(assignedConduitKey, owner.address);
-    });
+    await conduitController
+      .connect(owner)
+      .createConduit(assignedConduitKey, owner.address);
 
     const tempConduit = conduitImplementation.attach(tempConduitAddress);
     return tempConduit;
