@@ -1,33 +1,40 @@
 import { expect } from "chai";
-import { constants, Wallet } from "ethers";
+import { constants } from "ethers";
 import { keccak256, recoverAddress } from "ethers/lib/utils";
 import hre, { ethers } from "hardhat";
-import {
-  ConduitInterface,
-  ConduitControllerInterface,
-  ImmutableCreate2FactoryInterface,
-  ConsiderationInterface,
-  TestZone,
-} from "../../../typechain-types";
+
 import { deployContract } from "../contracts";
+import { getBulkOrderTree } from "../eip712/bulk-orders";
 import {
   calculateOrderHash,
   convertSignatureToEIP2098,
   randomHex,
   toBN,
 } from "../encoding";
-import {
+import { VERSION } from "../helpers";
+
+import type {
+  ConduitControllerInterface,
+  ConduitInterface,
+  ConsiderationInterface,
+  ImmutableCreate2FactoryInterface,
+  TestInvalidContractOfferer,
+  TestInvalidContractOffererRatifyOrder,
+  TestPostExecution,
+  TestZone,
+} from "../../../typechain-types";
+import type {
   AdvancedOrder,
   ConsiderationItem,
   CriteriaResolver,
   OfferItem,
   OrderComponents,
 } from "../types";
+import type { Contract, Wallet } from "ethers";
 
-const { orderType } = require("../../../eip-712-types/order");
 const deployConstants = require("../../../constants/constants");
-
-const VERSION = !process.env.REFERENCE ? "1.1" : "rc.1.1";
+// const { bulkOrderType } = require("../../../eip-712-types/bulkOrder");
+const { orderType } = require("../../../eip-712-types/order");
 
 export const marketplaceFixture = async (
   create2Factory: ImmutableCreate2FactoryInterface,
@@ -41,11 +48,12 @@ export const marketplaceFixture = async (
     process.env.REFERENCE ? "ReferenceConsideration" : "Seaport"
   );
 
-  const directMarketplaceContract = await deployContract(
-    process.env.REFERENCE ? "ReferenceConsideration" : "Consideration",
-    owner as any,
-    conduitController.address
-  );
+  const directMarketplaceContract =
+    await deployContract<ConsiderationInterface>(
+      process.env.REFERENCE ? "ReferenceConsideration" : "Consideration",
+      owner,
+      conduitController.address
+    );
 
   const marketplaceContractAddress = await create2Factory.findCreate2Address(
     deployConstants.MARKETPLACE_CONTRACT_CREATION_SALT,
@@ -78,20 +86,36 @@ export const marketplaceFixture = async (
     .connect(owner)
     .updateChannel(conduitOne.address, marketplaceContract.address, true);
 
-  const stubZone: TestZone = await deployContract("TestZone", owner as any);
+  const stubZone = await deployContract<TestZone>("TestZone", owner);
+  const postExecutionZone = await deployContract<TestPostExecution>(
+    "TestPostExecution",
+    owner
+  );
+
+  const invalidContractOfferer =
+    await deployContract<TestInvalidContractOfferer>(
+      "TestInvalidContractOfferer",
+      owner,
+      marketplaceContractAddress
+    );
+
+  const invalidContractOffererRatifyOrder =
+    await deployContract<TestInvalidContractOffererRatifyOrder>(
+      "TestInvalidContractOffererRatifyOrder",
+      owner,
+      marketplaceContractAddress
+    );
 
   // Required for EIP712 signing
   const domainData = {
     name: process.env.REFERENCE ? "Consideration" : "Seaport",
     version: VERSION,
-    chainId: chainId,
+    chainId,
     verifyingContract: marketplaceContract.address,
   };
 
   const getAndVerifyOrderHash = async (orderComponents: OrderComponents) => {
-    const orderHash = await marketplaceContract.getOrderHash(
-      orderComponents as any
-    );
+    const orderHash = await marketplaceContract.getOrderHash(orderComponents);
     const derivedOrderHash = calculateOrderHash(orderComponents);
     expect(orderHash).to.equal(derivedOrderHash);
     return orderHash;
@@ -100,7 +124,7 @@ export const marketplaceFixture = async (
   // Returns signature
   const signOrder = async (
     orderComponents: OrderComponents,
-    signer: Wallet
+    signer: Wallet | Contract
   ) => {
     const signature = await signer._signTypedData(
       domainData,
@@ -121,9 +145,65 @@ export const marketplaceFixture = async (
     return signature;
   };
 
+  const signBulkOrder = async (
+    orderComponents: OrderComponents[],
+    signer: Wallet | Contract,
+    startIndex = 0,
+    height?: number,
+    extraCheap?: boolean
+  ) => {
+    const tree = getBulkOrderTree(orderComponents, startIndex, height);
+    const bulkOrderType = tree.types;
+    const chunks = tree.getDataToSign();
+    let signature = await signer._signTypedData(domainData, bulkOrderType, {
+      tree: chunks,
+    });
+
+    if (extraCheap) {
+      signature = convertSignatureToEIP2098(signature);
+    }
+
+    const proofAndSignature = tree.getEncodedProofAndSignature(
+      startIndex,
+      signature
+    );
+
+    const orderHash = tree.getBulkOrderHash();
+
+    const { domainSeparator } = await marketplaceContract.information();
+    const digest = keccak256(
+      `0x1901${domainSeparator.slice(2)}${orderHash.slice(2)}`
+    );
+    const recoveredAddress = recoverAddress(digest, signature);
+
+    expect(recoveredAddress).to.equal(signer.address);
+
+    // Verify each individual order
+    for (const components of orderComponents) {
+      const individualOrderHash = await getAndVerifyOrderHash(components);
+      const digest = keccak256(
+        `0x1901${domainSeparator.slice(2)}${individualOrderHash.slice(2)}`
+      );
+      const individualOrderSignature = await signer._signTypedData(
+        domainData,
+        orderType,
+        components
+      );
+      const recoveredAddress = recoverAddress(digest, individualOrderSignature);
+      expect(recoveredAddress).to.equal(signer.address);
+    }
+
+    return proofAndSignature;
+  };
+
   const createOrder = async (
-    offerer: Wallet,
-    zone: Wallet | undefined | string = undefined,
+    offerer: Wallet | Contract,
+    zone:
+      | TestZone
+      | TestPostExecution
+      | Wallet
+      | undefined
+      | string = undefined,
     offer: OfferItem[],
     consideration: ConsiderationItem[],
     orderType: number,
@@ -132,7 +212,10 @@ export const marketplaceFixture = async (
     signer?: Wallet,
     zoneHash = constants.HashZero,
     conduitKey = constants.HashZero,
-    extraCheap = false
+    extraCheap = false,
+    useBulkSignature = false,
+    bulkSignatureIndex?: number,
+    bulkSignatureHeight?: number
   ) => {
     const counter = await marketplaceContract.getCounter(offerer.address);
 
@@ -145,7 +228,7 @@ export const marketplaceFixture = async (
     const orderParameters = {
       offerer: offerer.address,
       zone: !extraCheap
-        ? (zone as Wallet).address || (zone as string)
+        ? (zone as Wallet).address ?? zone
         : constants.AddressZero,
       offer,
       consideration,
@@ -177,7 +260,7 @@ export const marketplaceFixture = async (
       totalSize,
     };
 
-    const flatSig = await signOrder(orderComponents, signer || offerer);
+    const flatSig = await signOrder(orderComponents, signer ?? offerer);
 
     const order = {
       parameters: orderParameters,
@@ -186,6 +269,28 @@ export const marketplaceFixture = async (
       denominator: 1, // only used for advanced orders
       extraData: "0x", // only used for advanced orders
     };
+
+    if (useBulkSignature) {
+      order.signature = await signBulkOrder(
+        [orderComponents],
+        signer ?? offerer,
+        bulkSignatureIndex,
+        bulkSignatureHeight,
+        extraCheap
+      );
+
+      // Verify bulk signature length
+      expect(
+        order.signature.slice(2).length / 2,
+        "bulk signature length should be valid (98 < length < 837)"
+      )
+        .to.be.gt(98)
+        .and.lt(837);
+      expect(
+        (order.signature.slice(2).length / 2 - 67) % 32,
+        "bulk signature length should be valid ((length - 67) % 32 < 2)"
+      ).to.be.lt(2);
+    }
 
     // How much ether (at most) needs to be supplied when fulfilling the order
     const value = offer
@@ -215,6 +320,8 @@ export const marketplaceFixture = async (
       value,
       orderStatus,
       orderComponents,
+      startTime,
+      endTime,
     };
   };
 
@@ -433,9 +540,9 @@ export const marketplaceFixture = async (
       counter,
     };
 
-    const flatSig = await signOrder(orderComponents as any, offerer);
+    const flatSig = await signOrder(orderComponents, offerer);
 
-    const mirrorOrderHash = await getAndVerifyOrderHash(orderComponents as any);
+    const mirrorOrderHash = await getAndVerifyOrderHash(orderComponents);
 
     const mirrorOrder = {
       parameters: orderParameters,
@@ -467,8 +574,12 @@ export const marketplaceFixture = async (
     marketplaceContract,
     directMarketplaceContract,
     stubZone,
+    postExecutionZone,
+    invalidContractOfferer,
+    invalidContractOffererRatifyOrder,
     domainData,
     signOrder,
+    signBulkOrder,
     createOrder,
     createMirrorBuyNowOrder,
     createMirrorAcceptOfferOrder,
